pyproject.toml
[project]
name = "11x-rag-example"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
    "zep-cloud>=2.4.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "ruff>=0.9.7",
]
